Restructure code
