--- TF/README.md
@@ -1,4 +1,0 @@
-# TF Models (.pb files)
-Enter the following commands to run Inferencing <br>
-- `bazel build main` <br>
-- `bazel-bin/main "filepath" "model_name" "category" "is_saved_model"` <br>
