--- TFLite/README.md
@@ -1,6 +1,0 @@
-# TFLite Models (.tflite files)
-Enter the following commands to run Inferencing <br>
-- `bazel build init` <br>
-- `bazel-bin/init` <br>
-- `bazel build main` <br>
-- `bazel-bin/main "filepath" "model_name" "category" ` <br>
